# api-client/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Decorating Platform API",
    "description" : "Quoting and customer modules.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Same origin"
  } ],
...
  "paths" : {
...
    "/api/otp/verify" : {
      "post" : {
        "tags" : [ "Verification" ],
        "summary" : "Verify the code, creating the customer",
        "operationId" : "verifyCode",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VerifyOtpRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Verified"
          },
          "401" : {
            "description" : "No session cookie"
          },
          "422" : {
            "description" : "Wrong, expired or already used"
          },
          "423" : {
            "description" : "§11's fifth wrong guess"
          }
        }
      }
    },
    "/api/otp/send" : {
      "post" : {
        "tags" : [ "Verification" ],
        "summary" : "Send a verification code to a number",
        "operationId" : "sendCode",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SendOtpRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Queued — see §13 on what 'sent' means here — with the code's expiry",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendOtpResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Not a Turkish mobile number",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendOtpResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "No session cookie",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendOtpResponse"
                }
              }
            }
          },
          "429" : {
            "description" : "§11's limits",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendOtpResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "VerifyOtpRequest" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "example" : "123456",
            "minLength" : 1
          }
        },
        "required" : [ "code" ]
      },
      "SendOtpRequest" : {
        "type" : "object",
        "properties" : {
          "phone" : {
            "type" : "string",
            "example" : "0555 123 45 67",
            "minLength" : 1
          }
        },
        "required" : [ "phone" ]
      },
      "SendOtpResponse" : {
        "type" : "object",
        "properties" : {
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "expiresAt" ]
      },
...
    }
  }
}